+++ pan-tilt-cli/src/system.h
@@ -17,7 +17,7 @@
 
 //// Defines //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-#define SYSTEM_VERSION "1.6.0"
+#define SYSTEM_VERSION "1.7.0"
 
 //// Public Declarations [Interface] //////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -36,17 +36,21 @@ namespace sys
 	// public methods
 
 	void connect(std::string com_port);
+	
 	void gui();
+	void stream_handler(uart::UART_FRAME frame);
 
 	void write_byte(std::string byte);
 	void write_array(std::string args);
 	void write_spi(std::string args);
 
 	void echo();
-	void set_mode(std::string& args);
-	void set_pwm(std::string& args);
-	void set_freq(std::string& args);
+	void set_mode(std::string args);
+	void set_gui(std::string args);
+	void set_msg(std::string args);
+	void set_pwm(std::string args);
+	void set_freq(std::string args);
 
-	void get_enc(std::string& args);
-	void get_hal(std::string& args);
+	void get_enc(std::string args);
+	void get_hal(std::string args);
 }